refactor(reservation-domain): Room 패키지로 도메인 분리
- Room, RoomOperationPolicy, 관련 Enum 대상으로 domain.room 패키지로 이동함
- Test, TestFixture 도 패키지 맞춤
