+++ reservation/reservation-application/src/test/java/com/seatliberator/seatliberator/reservation/availability/application/service/FindAvailableSeatsUseCaseTest.java
@@ -7,7 +7,7 @@ import com.seatliberator.seatliberator.reservation.application.availability.serv
 import com.seatliberator.seatliberator.reservation.application.booking.contract.OccupancySeatLocatorFinder;
 import com.seatliberator.seatliberator.reservation.application.booking.contract.OccupancySeatRangeFinder;
 import com.seatliberator.seatliberator.reservation.application.room.port.out.SeatReader;
-import com.seatliberator.seatliberator.reservation.domain.fixture.RoomFixture;
+import com.seatliberator.seatliberator.reservation.domain.room.RoomFixture;
 import com.seatliberator.seatliberator.reservation.domain.seat.SeatFixture;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
