+++ reservation/reservation-application/src/test/java/com/seatliberator/seatliberator/reservation/room/application/RoomQueryServiceTest.java
@@ -18,7 +18,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import java.util.List;
 import java.util.Optional;
 
-import static com.seatliberator.seatliberator.reservation.domain.fixture.RoomFixture.createRoom;
+import static com.seatliberator.seatliberator.reservation.domain.room.RoomFixture.createRoom;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.Mockito.verify;
