+++ reservation/reservation-application/src/test/java/com/seatliberator/seatliberator/reservation/room/application/SeatQueryServiceTest.java
@@ -22,8 +22,8 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import java.util.List;
 import java.util.Optional;
 
-import static com.seatliberator.seatliberator.reservation.domain.fixture.RoomFixture.createRoom;
 import static com.seatliberator.seatliberator.reservation.domain.fixture.TestSupport.fixedClock;
+import static com.seatliberator.seatliberator.reservation.domain.room.RoomFixture.createRoom;
 import static com.seatliberator.seatliberator.reservation.domain.seat.SeatFixture.create;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
