--- reservation/reservation-domain/src/main/java/com/seatliberator/seatliberator/reservation/domain/persistence/RoomOperationStatus.java
@@ -1,6 +1,0 @@
-package com.seatliberator.seatliberator.reservation.domain.persistence;
-
-public enum RoomOperationStatus {
-    OPEN,
-    CLOSE
-}
