+++ reservation/reservation-domain/src/main/java/com/seatliberator/seatliberator/reservation/domain/room/RoomOperationStatus.java
@@ -1,0 +1,6 @@
+package com.seatliberator.seatliberator.reservation.domain.room;
+
+public enum RoomOperationStatus {
+    OPEN,
+    CLOSE
+}
