+++ reservation/reservation-domain/src/main/java/com/seatliberator/seatliberator/reservation/domain/seat/Seat.java
@@ -3,7 +3,7 @@ package com.seatliberator.seatliberator.reservation.domain.seat;
 import com.seatliberator.seatliberator.kernel.condition.Preconditions;
 import com.seatliberator.seatliberator.reservation.domain.SeatLocator;
 import com.seatliberator.seatliberator.reservation.domain.SimpleSeatLocator;
-import com.seatliberator.seatliberator.reservation.domain.persistence.Room;
+import com.seatliberator.seatliberator.reservation.domain.room.Room;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
